+++ e2e/e2etests/test_bitcoin_withdraw.go
@@ -4,22 +4,27 @@ import (
 	"fmt"
 	"math/big"
 
+	"github.com/btcsuite/btcd/btcjson"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcutil"
+	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/e2e/runner"
 	"github.com/zeta-chain/zetacore/e2e/utils"
+	"github.com/zeta-chain/zetacore/zetaclient/testutils"
 )
 
 func TestBitcoinWithdraw(r *runner.E2ERunner) {
-	// withdraw 0.1 BTC from ZRC20 to BTC address
-	// first, approve the ZRC20 contract to spend 1 BTC from the deployer address
+	// withdraw 0.01 BTC from ZRC20 to BTC address
 	WithdrawBitcoin(r)
 }
 
-func WithdrawBitcoin(r *runner.E2ERunner) {
-	amount := big.NewInt(0.1 * btcutil.SatoshiPerBitcoin)
+func TestBitcoinWithdrawRestricted(r *runner.E2ERunner) {
+	// withdraw 0.01 BTC from ZRC20 to BTC restricted address
+	WithdrawBitcoinRestricted(r)
+}
 
-	// approve the ZRC20 contract to spend 1 BTC from the deployer address
+func withdrawBTCZRC20(r *runner.E2ERunner, to btcutil.Address, amount *big.Int) *btcjson.TxRawResult {
+	// approve the ZRC20 contract to spend 'amount' of BTC from the deployer address
 	tx, err := r.BTCZRC20.Approve(r.ZevmAuth, r.BTCZRC20Addr, big.NewInt(amount.Int64()*2)) // approve more to cover withdraw fee
 	if err != nil {
 		panic(err)
@@ -32,8 +37,8 @@ func WithdrawBitcoin(r *runner.E2ERunner) {
 	// mine blocks
 	stop := r.MineBlocks()
 
-	// withdraw 0.1 BTC from ZRC20 to BTC address
-	tx, err = r.BTCZRC20.Withdraw(r.ZevmAuth, []byte(r.BTCDeployerAddress.EncodeAddress()), amount)
+	// withdraw 'amount' of BTC from ZRC20 to BTC address
+	tx, err = r.BTCZRC20.Withdraw(r.ZevmAuth, []byte(to.EncodeAddress()), amount)
 	if err != nil {
 		panic(err)
 	}
@@ -43,7 +48,7 @@ func WithdrawBitcoin(r *runner.E2ERunner) {
 	}
 
 	// mine 10 blocks to confirm the withdraw tx
-	_, err = r.BtcRPCClient.GenerateToAddress(10, r.BTCDeployerAddress, nil)
+	_, err = r.BtcRPCClient.GenerateToAddress(10, to, nil)
 	if err != nil {
 		panic(err)
 	}
@@ -75,6 +80,29 @@ func WithdrawBitcoin(r *runner.E2ERunner) {
 
 	// stop mining
 	stop <- struct{}{}
+
+	return rawTx
+}
+
+func WithdrawBitcoin(r *runner.E2ERunner) {
+	amount := big.NewInt(0.01 * btcutil.SatoshiPerBitcoin)
+	withdrawBTCZRC20(r, r.BTCDeployerAddress, amount)
+}
+
+func WithdrawBitcoinRestricted(r *runner.E2ERunner) {
+	amount := big.NewInt(0.01 * btcutil.SatoshiPerBitcoin)
+
+	// use restricted BTC P2WPKH address
+	addressRestricted, err := common.DecodeBtcAddress(testutils.RestrictedBtcAddressTest, common.BtcRegtestChain().ChainId)
+	if err != nil {
+		panic(err)
+	}
+
+	// the cctx should be cancelled
+	rawTx := withdrawBTCZRC20(r, addressRestricted, amount)
+	if len(rawTx.Vout) != 2 {
+		panic(fmt.Errorf("BTC cancelled outtx rawTx.Vout should have 2 outputs"))
+	}
 }
 
 // WithdrawBitcoinMultipleTimes ...